integer

+++ app/models.py
@@ -45,7 +45,7 @@ class Mechanics(Base):
     name: Mapped[str] = mapped_column(db.String(255), nullable=False)
     email: Mapped[str] = mapped_column(db.String(255), nullable=False, unique=True)
     phone_number: Mapped[str] = mapped_column(db.String(255), nullable=False)
-    salary: Mapped[float] = mapped_column(db.Float, nullable=False)
+    salary: Mapped[float] = mapped_column(db.Integer, nullable=False)
 
     service_tickets: Mapped[List['Service_Tickets']] = relationship(secondary=service_mechanics, back_populates='mechanics')
